add TypeScript types to app/community and app/meals files; remove old .js files after Typescript migration
