--- app/meals/error.js
@@ -1,10 +1,0 @@
-'use client';
-
-export default function Error() {
-    return (
-        <main className="error">
-            <h1>An error occurred!</h1>
-            <p>Failed to fetch meal data. Please try again later.</p>
-        </main>
-    );
-}
