+++ app/meals/not-found.tsx
@@ -1,0 +1,8 @@
+export default function NotFound() {
+  return (
+    <main className="not-found">
+      <h1>Meal not found</h1>
+      <p>Unfortunately, we could not find the requested page or meal data.</p>
+    </main>
+  );
+}
