+++ app/meals/share/page.tsx
@@ -4,10 +4,11 @@ import ImagePicker from '@/components/meals/image-picker';
 import { useFormState } from 'react-dom';
 import { shareMeal } from '@/lib/actions';
 import MealsFormSubmit from '@/components/meals/meals-form-submit';
+import type { FormState } from '@/components/meals/types';
 import styles from './page.module.scss';
 
 export default function ShareMealPage() {
-  const [state, formAction] = useFormState(shareMeal, { message: null });
+  const [state, formAction] = useFormState<FormState, FormData>(shareMeal, { message: null });
 
   return (
     <>
@@ -42,7 +43,7 @@ export default function ShareMealPage() {
             <textarea
               id="instructions"
               name="instructions"
-              rows="10"
+              rows={10}
               required
             ></textarea>
           </p>
